Add Create Pet and Change Pet functionality
Allows for creating pet with a name (checked for uniqueness) and
specifying a type of pet. Creating a pet writes to a file named
"profile" which contains a list of every pet. Creating a pet will also
create a file whose name is the name of the pet.

When changing pets, the app dynamically creates radio buttosn on a
dialog fragment that correspond to each saved pet's name. Selecting a
pet from this list will reload the UI with the chosen pet's details.
Users can also add pets from this dialog.

When logging into the app, the user is prompted to select a pet if one
exists. Otherwise, the user is prompted immediately to create their
first pet.