Update behavior for navigating between pages.
When launching the app, the user is no longer prompted to select a pet.
Instead, the first pet is selected by default. The user can also tap on
a card on the home screen to select a pet and immediately navigate to
the corresponding profile.

The cards on the home page now display name "the" type (e.g. Megan the
Poodle). This change resulted in combining the name and type into a
single object saved in the JSON, both for the list of pets and for
individual pet profiles.

Additionally, when leaving the profile page, the app will check for
unsaved changes. The user has the option to either save, discard, or
cancel. This check is also performed when attempting to select a
different pet or a dd a new one.

Improve some general layout formatting.