Add methods to delete profiles and fix low quality images.
We can now delete profiles.

Icons can now get images on the home page. All ImageViews have been
replaced with a circular ImageView from an external library.

Low quality photos are a thing of the past! There is now a loading
indicator that appears when setting pictures. This was introduced
particularly because there is a delay when setting from the camera for
the first time. The screen becomes half-transparent during the load.

Picture export is functional.